Run iceberg tests as a separate suite.
Not all tests need to run as iceberg. Instead only run the SQL tests as
iceberg in addition to running them as CSV to test base behavior.
